requirements.txt
gTTS==2.5.3
numpy==2.3.0
openai==1.86.0
pvporcupine==3.0.3
pyaudio==0.2.14
pyautogui==0.9.54
python-dotenv==1.1.0
python_vlc==3.0.21203
requests==2.32.4
scipy==1.15.3
sounddevice==0.5.1
SpeechRecognition==3.11.0
yt_dlp==2025.6.9
